+++ Carthage/Checkouts/ReactiveReSwift/Sources/CoreTypes/Store.swift
@@ -17,37 +17,27 @@ import Foundation
 
 import Foundation
 
-public class Store<ObservableProperty: ObservablePropertyType>: StoreType where ObservableProperty.ValueType: StateType {
+open class Store<ObservableProperty: ObservablePropertyType> where ObservableProperty.ValueType: StateType {
 
     public typealias StoreMiddleware = Middleware<ObservableProperty.ValueType>
     public typealias StoreReducer = Reducer<ObservableProperty.ValueType>
 
     public var dispatchMiddleware: Middleware<ObservableProperty.ValueType>!
-
     private var reducer: StoreReducer
-
     public var observable: ObservableProperty!
-
-    private let dispatchQueue: DispatchQueue
-
     private var disposeBag = SubscriptionReferenceBag()
 
     public required init(reducer: StoreReducer,
-                         stateType: ObservableProperty.ValueType.Type,
                          observable: ObservableProperty,
-                         middleware: StoreMiddleware = Middleware(),
-                         dispatchQueue: DispatchQueue = DispatchQueue.main) {
+                         middleware: StoreMiddleware = Middleware()) {
         self.reducer = reducer
         self.observable = observable
         self.dispatchMiddleware = middleware
-        self.dispatchQueue = dispatchQueue
     }
 
     private func defaultDispatch(action: Action) {
-        dispatchQueue.sync {
-            let value = self.reducer.transform(action, self.observable.value)
-            self.observable.value = value
-        }
+        let value = self.reducer.transform(action, self.observable.value)
+        self.observable.value = value
     }
 
     public func dispatch(_ actions: Action...) {