+++ Carthage/Checkouts/ReactiveReSwift/Tests/Observable/StoreTests.swift
@@ -19,7 +19,6 @@ class ObservableStoreTests: XCTestCase {
 
         autoreleasepool {
             _ = DeInitStore(reducer: testReducer,
-                                      stateType: TestAppState.self,
                                       observable: ObservableProperty(TestAppState()),
                                       deInitAction: { deInitCount += 1 })
         }
@@ -37,29 +36,21 @@ class DeInitStore<State: StateType>: Store<ObservableProperty<State>> {
         deInitAction?()
     }
 
-    convenience init(reducer: Reducer<ObservableProperty.ValueType>,
-                  stateType: ObservableProperty.ValueType.Type,
-                  observable: ObservableProperty,
-                  middleware: Middleware<ObservableProperty.ValueType> = Middleware(),
-                  dispatchQueue: DispatchQueue = DispatchQueue.main,
+    convenience init(reducer: Reducer<ObservableProperty<State>.ValueType>,
+                  observable: ObservableProperty<State>,
+                  middleware: Middleware<ObservableProperty<State>.ValueType> = Middleware(),
                   deInitAction: @escaping () -> Void) {
         self.init(reducer: reducer,
-                   stateType: stateType,
                    observable: observable,
-                   middleware: middleware,
-                   dispatchQueue: dispatchQueue)
+                   middleware: middleware)
         self.deInitAction = deInitAction
     }
 
-    required init(reducer: Reducer<ObservableProperty.ValueType>,
-                  stateType: ObservableProperty.ValueType.Type,
-                  observable: ObservableProperty,
-                  middleware: Middleware<ObservableProperty.ValueType>,
-                  dispatchQueue: DispatchQueue) {
+    required init(reducer: Reducer<ObservableProperty<State>.ValueType>,
+                  observable: ObservableProperty<State>,
+                  middleware: Middleware<ObservableProperty<State>.ValueType>) {
         super.init(reducer: reducer,
-                   stateType: stateType,
                    observable: observable,
-                   middleware: middleware,
-                   dispatchQueue: dispatchQueue)
+                   middleware: middleware)
     }
 }